refactor folder and package structures
