+++ apps/react/public/index.html
@@ -5,12 +5,9 @@
         <link href="favicon.png" rel="icon">
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <title>React app</title>
-
         <link rel="modulepreload" href="/cdn/movie-widgets/index.js" />
-        <script type="module" src="/cdn/movie-widgets/index.js" async="false"></script>
-        <link rel="modulepreload" href="/cdn/workleap-header/index.js" />
-        <script type="module" src="/cdn/workleap-header/index.js" async="false"></script>
-        <script type="module" src="/index.js" blocking="render" async="false"></script>
+        <script type="module" src="/cdn/movie-widgets/index.js"></script>
+        <script type="module" src="index.js" blocking="render" async="false"></script>
 
     </head>
     <body>
