+++ apps/react/src/Layout.tsx
@@ -1,8 +1,5 @@
-// apps/react/src/Layout.tsx
 import { Button, Div, Flex } from "@workleap/orbiter-ui";
-import React from "react";
 import { Link, Outlet } from "react-router-dom";
-import { WorkleapHeader } from "workleap-nav/react";
 import { MovieFinder, SelectedMovie } from "../widgets/Widgets.tsx";
 import { AppContextProvider, useAppContext } from "./AppContext.tsx";
 
@@ -13,8 +10,6 @@ export function Layout() {
     return (
         <AppContextProvider>
             <Div>
-                <WorkleapHeader data={{}} >
-                </WorkleapHeader>
                 <header style={{ borderBottom: "1px solid black", padding: "5px", display: "flex", justifyContent: "space-between" }}>
                     <div>
                         <h1>React App</h1>
