+++ apps/react/webpack.dev.js
@@ -18,9 +18,6 @@ function virtualCdn(config) {
         }, {
             publicPath: "/cdn/movie-widgets",
             directory: join(__dirname, "../../widgets/dist")
-        }, {
-            publicPath: "/cdn/workleap-header",
-            directory: "/home/mahmoudmoravej/workleap/workleap-navigation-components/dist/cdn"
         }]
     };
