+++ apps/vanilla-js/public/index.js
@@ -1,10 +1,6 @@
 import { MovieWidgets } from "/cdn/movie-widgets/index.js";
-import { WorkleapHeaderManager } from "/cdn/workleap-header/index.js";
-
 
 MovieWidgets.initialize({ theme: "light" });
-WorkleapHeaderManager.initialize();
-
 
 (function init() {
     handleMovieDetailsClick();
