+++ packages/movie-widgets/.eslintignore
