+++ packages/movie-widgets/package.json
@@ -1,10 +1,24 @@
 {
-    "name": "@workleap-movie-sample/web-components",
+    "name": "workleap-movie",
+    "version": "1.0.0",
     "author": "Workleap",
     "type": "module",
     "private": true,
+    "main": "dist/react/index.js",
+    "types": "dist/react/index.d.ts",
+    "exports": {
+        "./react": {
+            "import": "./dist/react/index.js",
+            "types": "./dist/react/index.d.ts"
+        }
+    },
+    "files": [
+        "dist"
+    ],
     "scripts": {
-        "build": "tsup --config ./tsup.build.ts",
+        "build": "pnpm run \"/^build:.*/\"",
+        "build:cdn": "tsup --config ./tsup.build.cdn.ts",
+        "build:react": "tsup --config ./tsup.build.react.ts",
         "lint": "pnpm run \"/^lint:.*/\"",
         "lint:js": "eslint .",
         "lint:types": "tsc --noEmit"
@@ -15,7 +29,7 @@
         "@workleap/orbiter-ui": "5.5.0",
         "react": "18.3.1",
         "react-dom": "18.3.1",
-        "@workleap/r2wc": "file:/home/mahmoudmoravej/workleap/wl-framework-agnostic-widgets-poc/packages/r2wc/workleap-r2wc-1.0.0.tgz"
+        "@workleap/r2wc": "workspace:./packages/r2wc/"
     },
     "devDependencies": {
         "@types/react": "18.3.11",
