+++ packages/movie-widgets/src/MovieDetails.tsx
