+++ packages/movie-widgets/src/MovieFinderElement.tsx
@@ -1,5 +1,5 @@
-import { WebComponentHTMLElement } from "@workleap/r2wc/core";
-import { MovieFinder } from "../react/MovieFinder.tsx";
+import { WebComponentHTMLElement } from "@workleap/r2wc";
+import { MovieFinder } from "./MovieFinder.tsx";
 
 export class MovieFinderElement extends WebComponentHTMLElement {
     get reactComponent() {
