+++ packages/movie-widgets/src/SelectedMovie.tsx
