+++ packages/movie-widgets/src/Ticket.tsx
