+++ packages/movie-widgets/src/index.ts
@@ -1,0 +1,4 @@
+import "./main.css";
+
+export * from "./widgets.ts";
+
