+++ packages/movie-widgets/tsup.build.cdn.ts
@@ -2,8 +2,9 @@ import { defineBuildConfig } from "@workleap/tsup-configs";
 
 export default defineBuildConfig({
     entry: ["src/index.ts"],
-    outDir: "dist",
+    outDir: "dist/cdn",
     noExternal: [/.*/],
+    dts: false,
     minify: true,
     treeshake: true
 });
