+++ packages/movie-widgets/tsup.build.react.ts
@@ -1,8 +1,9 @@
 import { defineBuildConfig } from "@workleap/tsup-configs";
 
 export default defineBuildConfig({
-    entry: [ "src/index.ts", "src/helpers/react/index.ts" ],
-    outDir: "dist",
+    entry: ["src/helpers/react/index.ts"],
+    outDir: "dist/react",
+    // tsconfig: "tsconfig.build.react.json",
     dts: true,
     clean: true,
     sourcemap: true
