--- packages/r2wc/src/core/index.ts
@@ -1,4 +1,0 @@
-export { type Map } from "./utils.ts";
-export { WebComponentHTMLElement, WebComponentHTMLElementBase } from "./WebComponentHTMLElement.tsx";
-export { WidgetsManager } from "./WidgetsManager.tsx";
-
