--- widgets/src/index.ts
@@ -1,4 +1,0 @@
-import "./main.css";
-
-export * from "./web-components/widgets.ts";
-
